requirements.txt
ultralytics
torch
pillow
numpy
opencv-python
